pyproject.toml
[project]
name = "mcp-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "ruff>=0.9.10",
    "starlette>=0.34.0",
    "uvicorn>=0.25.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.6.0",
]

[tool.ruff]
line-length = 100
